# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "hw7-swagger",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "Домашня робота №7"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "tags": [
    {
      "name": "Contacts",
      "description": "Operation with contact"
    }
  ],
  "paths": {
    "/contacts": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Get all contacts",
        "operationId": "getAllContacts",
        "description": "Get all contacts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "name": "perPage",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 10
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "example": "asc"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "example": "contactType"
            }
          },
          {
            "name": "isFavourite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "example": false
            }
          },
          {
            "name": "contactType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["home", "work", "personal"],
              "example": "home"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get all contacts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found contacts!"
                    },
                    "data": {
                      "$ref": "#/components/schemas/contact"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "tags": ["Contacts"],
        "summary": "Created contact",
        "operationId": "createContact",
        "description": "Create contact",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "phoneNumber",
                  "isFavourite",
                  "contactType"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Contact name",
                    "example": "Nejk dima"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "description": "Contact phone",
                    "example": 3802341230
                  },
                  "isFavourite": {
                    "type": "boolean",
                    "description": "Contact favourite",
                    "example": false
                  },
                  "contactType": {
                    "type": "string",
                    "description": "Contact type",
                    "enum": ["work", "home", "personal"],
                    "example": "personal"
                  },
                  "email": {
                    "type": "string",
                    "description": "Email contact",
                    "example": "doe@mail.com",
                    "default": false
                  },
                  "photo": {
                    "type": "string",
                    "format": "binary",
                    "description": "Contact photo",
                    "default": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create a contact",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Contacts create contacts"
                    },
                    "data": {
                      "$ref": "#/components/schemas/contact"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Get contact by id",
        "operationId": "getContacts",
        "description": "Get contact by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "6841738d9abba11274e4c270"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get contact by id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully get contact"
                    },
                    "data": {
                      "$ref": "#/components/schemas/contact"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "tags": ["Contacts"],
        "summary": "Delete contacts",
        "operationId": "deleteContact",
        "description": "Delete contacts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "6841738d9abba11274e4c270"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Contacts delete"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "patch": {
        "tags": ["Contacts"],
        "summary": "Update contacts",
        "operationId": "updateContact",
        "description": "Update contacts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "6841738d9abba11274e4c270"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Contact name",
                    "example": "Nejk dima"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "description": "Phone contact",
                    "example": 3802341230
                  },
                  "email": {
                    "type": "string",
                    "description": "Email contact",
                    "example": "doe@mail.com"
                  },
                  "isFavourite": {
                    "type": "boolean",
                    "description": "Contact favourite",
                    "example": false
                  },
                  "contactType": {
                    "type": "string",
                    "description": "Contact type",
                    "enum": ["work", "home", "personal"],
                    "example": "personal"
                  },
                  "photo": {
                    "type": "string",
                    "format": "url",
                    "description": "Contact photo url",
                    "example": "https://res.cloudinary.com/df5xsxnzm/image/upload/v1749232498/jpolliyga4egrctz965w.jpg"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contact not update",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Contact not update"
                    },
                    "data": {
                      "$ref": "#/components/schemas/contact"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": "Contact update",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Contact update"
                    },
                    "data": {
                      "$ref": "#/components/schemas/contact"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register user",
        "operationId": "registerUser",
        "description": "Register user",
        "security": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password", "timestamps"],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 16,
                    "description": "User name",
                    "example": "Nejk dima"
                  },
                  "email": {
                    "type": "string",
                    "description": "Email user",
                    "example": "doe@mail.com"
                  },
                  "password": {
                    "type": "string",
                    "description": "User password",
                    "example": "doe@mail.com"
                  },
                  "timestamps": {
                    "type": "object",
                    "format": "date-time",
                    "description": "User created",
                    "example": "2025-06-06T17:54:59.164+00:00 2025-06-06T17:54:59.164+00:00"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully user created"
                    },
                    "data": {
                      "$ref": "#/components/schemas/user"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login user",
        "operationId": "loginUser",
        "description": "Login user",
        "security": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["password", "email"],
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "User email",
                    "example": "doe@mail.com"
                  },
                  "password": {
                    "type": "string",
                    "description": "User password",
                    "example": "1233qwe"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User login",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "user login"
                    },
                    "data": {
                      "$ref": "#/components/schemas/user"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logout user",
        "operationId": "logoutUser",
        "description": "logout user",
        "security": [],
        "responses": {
          "204": {
            "description": "User logout"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "contact": {
        "type": "object",
        "required": [
          "_id",
          "name",
          "phoneNumber",
          "email",
          "isFavourite",
          "contactType",
          "userId",
          "photo",
          "timestamps"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "Contact id",
            "example": "6841738d9abba11274e4c270"
          },
          "name": {
            "type": "string",
            "description": "Contact name",
            "example": "Nejk dima"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Phone contact",
            "example": 3802341230
          },
          "email": {
            "type": "string",
            "description": "Email contact",
            "example": "doe@mail.com"
          },
          "isFavourite": {
            "type": "boolean",
            "description": "Contact favourite",
            "example": false
          },
          "contactType": {
            "type": "string",
            "description": "Contact type",
            "enum": ["work", "home", "personal"],
            "example": "personal"
          },
          "userId": {
            "type": "string",
            "description": "Contact userId",
            "example": "684173599abba11274e4c268"
          },
          "photo": {
            "type": "string",
            "format": "url",
            "description": "Contact photo url",
            "example": "https://res.cloudinary.com/df5xsxnzm/image/upload/v1749232498/jpolliyga4egrctz965w.jpg"
          },
          "timestamps": {
            "type": "object",
            "format": "date-time",
            "description": "Contact created",
            "example": "2025-06-06T17:54:59.164+00:00 2025-06-06T17:54:59.164+00:00"
          }
        }
      },
      "user": {
        "type": "object",
        "required": ["_id", "name", "email", "timestamps"],
        "properties": {
          "_id": {
            "type": "string",
            "description": "User id",
            "example": "6841738d9abba11274e4c270"
          },
          "name": {
            "type": "string",
            "description": "User name",
            "example": "Nejk dima"
          },
          "email": {
            "type": "string",
            "description": "Email User",
            "example": "doe@mail.com"
          },
          "timestamps": {
            "type": "object",
            "format": "date-time",
            "description": "User created",
            "example": "2025-06-06T17:54:59.164+00:00 2025-06-06T17:54:59.164+00:00"
          }
        }
      }
    },
    "responses": {
      "401": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["status", "message", "data"],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "Unauthorized error"
                },
                "data": {
                  "type": "object",
                  "required": ["message"],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Email or password is incorrect"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["status", "message", "data"],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "Not found"
                },
                "data": {
                  "type": "object",
                  "required": ["message"],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Not found"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "409": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["status", "message", "data"],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 409
                },
                "message": {
                  "type": "string",
                  "example": "Email in use"
                },
                "data": {
                  "type": "object",
                  "required": ["message"],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Email in use"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "500": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["status", "message"],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "Internal server error"
                }
              }
            }
          }
        }
      }
    }
  }
}